footer relocation
