+++ src/components/layout/footer/footer.scss
@@ -1,5 +1,5 @@
 .footer-container {
-    margin: 50px 50px 100px 50px;
+    margin: 50px 32px 80px 32px;
     border-top: 1px solid var(--gold);
   }
 
@@ -72,9 +72,9 @@
     padding-bottom: 10px;
   }
 
-  @media screen and (min-width: 1024px) {
+  @media screen and (min-width: 767px) {
     .footer-container {
-        margin: 50px 100px 50px 100px;
+        margin: 50px 100px 80px 100px;
       }
     
       .footer-icons-container {
@@ -92,6 +92,7 @@
         margin-bottom: 20px;
         display: grid;
         grid-template-columns: repeat(4, 1fr);
+        gap: 20px;
       }
     
       .footer-bottom {
