+++ src/reset.scss
@@ -13,15 +13,14 @@
     --red: #D82700;
   }
 
-.container {   
-    min-width: 360px;   
-    margin: 0 auto;      
+  body {
+    border: 1px solid #000;
+    max-width: 1440px;
+    margin-right: auto;
+    margin-left: auto;
   }
 
-body {
-    min-width: 360px;
-    margin: 0 auto;  
-  }
+
